Corrected pointer type for work array in m_s_subsample.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/robustbase@315 694ef91d-65df-0310-b7bb-92e67a308ead

+++ pkg/robustbase/src/lmrob.c
@@ -2199,7 +2199,7 @@ void m_s_subsample(double *X1, double *y, int n, int p1, int p2,
     b_i =       (int *)    R_alloc(p2,    sizeof(int));
     ind_space = (int *)    R_alloc(n,     sizeof(int));
     xx =        (double *) R_alloc(p2*p2, sizeof(double));
-    work =      (double *) R_alloc(p2,    sizeof(int));
+    work =      (int *)    R_alloc(p2,    sizeof(int));
 
     /*	set the seed */
     GetRNGstate();
